STAT_TYPE instead of data_type in FactorAnalyser

+++ __init__.py
@@ -58,7 +58,7 @@ if 'SIDEKIT' in os.environ:
 
 PARALLEL_MODULE = 'multiprocessing'  # can be , threading, multiprocessing MPI is planned in the future
 PARAM_TYPE = numpy.float32
-STAT_TYPE = numpy.float32
+STAT_TYPE = numpy.float64  # can be numpy.float32 to speed up the computation but can lead to numerical issuess
 
 # Import bosaris-like classes
 from sidekit.bosaris import IdMap
